auth_service/requirements.txt
Django==5.2.7
djangorestframework
djangorestframework-simplejwt
kafka-python
kafka-python==2.1.0